+++ payway/constants.py
@@ -10,11 +10,9 @@ class PaymentMethod(StrEnum):
 
 
 PAYWAY_API_URL = "https://api.payway.com.au/rest/v1"
-TOKEN_URL = PAYWAY_API_URL + "/single-use-tokens-redirect"
+TOKEN_URL = PAYWAY_API_URL + "/single-use-tokens"
 TRANSACTION_URL = PAYWAY_API_URL + "/transactions"
 CUSTOMER_URL = PAYWAY_API_URL + "/customers"
-OWN_BANK_ACCOUNTS_URL = PAYWAY_API_URL + "/your-bank-accounts"
-TOKEN_NO_REDIRECT = PAYWAY_API_URL + "/single-use-tokens"
 TRANSACTION_APPROVED = "0"
 
 SUMMARY_CODES = {